refactor: 폴더명 복수로 변경
