+++ lib/features/promises/data/repositories/promises_repository_impl.dart
@@ -2,8 +2,8 @@ import 'package:dio/dio.dart' hide Headers;
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:pravo_client/features/core/dio/dio_provider.dart';
 import 'package:pravo_client/features/promises/data/models/promise_model.dart';
-import 'package:pravo_client/features/promises/domain/entity/promise.dart';
-import 'package:pravo_client/features/promises/domain/repository/promises_repository.dart';
+import 'package:pravo_client/features/promises/domain/entities/promise.dart';
+import 'package:pravo_client/features/promises/domain/repositories/promises_repository.dart';
 
 final promisesRepositoryProvider = Provider<PromisesRepository>((ref) {
   return PromisesRepositoryImpl(ref.read(dioProvider));
